+++ Changelog.md
@@ -28,6 +28,8 @@ Enhancements:
   ported from `rspec-fire` (Xavier Shay).
 * Improved performance of double creation, particularly those with many
   attributes. (Xavier Shay)
+* Default value of `transfer_nested_constants` option for constant stubbing can
+  be configured. (Xavier Shay)
 
 ### 2.14.3 / 2013-08-08
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.14.2...v2.14.3)