+++ Changelog.md
@@ -29,6 +29,11 @@ Enhancements:
 * Improved performance of double creation, particularly those with many
   attributes. (Xavier Shay)
 
+Bug Fixes:
+
+* Fix `and_call_original` to handle a complex edge case involving
+  singleton class ancestors. (Marc-André Lafortune, Myron Marston)
+
 ### 2.14.3 / 2013-08-08
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.14.2...v2.14.3)